利用規約を追加しました

+++ springboot-tabelog/src/main/java/com/example/springboot_tabelog/repository/TermRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.springboot_tabelog.entity.Term;
 
-public interface TermRepository extends JpaRepository<Term, Long> {
+public interface TermRepository extends JpaRepository<Term, Integer> {
 
 }
